Moves Ignorer over to lint/ignore.CmdIgnorer
