+++ pkg/lint/ignore/cmd_ignorer.go
@@ -1,9 +1,8 @@
-package lint
+package ignore
 
 import (
 	"bufio"
 	"fmt"
-	"helm.sh/helm/v3/pkg/lint/support"
 	"io"
 	"log"
 	"os"
@@ -11,22 +10,22 @@ import (
 	"strings"
 )
 
-// Ignorer provides a means of suppressing unwanted helm lint errors and messages
+// CmdIgnorer provides a means of suppressing unwanted helm lint errors and messages
 // by comparing them to an ignore list provided in a plaintext helm lint ignore file.
-type Ignorer struct {
+type CmdIgnorer struct {
 	Patterns        map[string][]string
 	ErrorPatterns   map[string][]string
 	debugFnOverride func(string, ...interface{})
 }
 
 // DefaultIgnoreFileName is the name of the lint ignore file
-// an Ignorer will seek out at load/parse time.
+// an CmdIgnorer will seek out at load/parse time.
 const DefaultIgnoreFileName = ".helmlintignore"
 
-// NewIgnorer builds an Ignorer object that enables helm to discard specific lint result Messages
+// NewIgnorer builds an CmdIgnorer object that enables helm to discard specific lint result Messages
 // and Errors should they match the ignore rules in the specified .helmlintignore file.
-func NewIgnorer(chartPath, ignoreFilePath string, debugLogFn func(string, ...interface{})) *Ignorer {
-	ignorer := &Ignorer{
+func NewIgnorer(chartPath, ignoreFilePath string, debugLogFn func(string, ...interface{})) *CmdIgnorer {
+	ignorer := &CmdIgnorer{
 		debugFnOverride: debugLogFn,
 		Patterns:        make(map[string][]string),
 		ErrorPatterns:   make(map[string][]string),
@@ -42,39 +41,11 @@ func NewIgnorer(chartPath, ignoreFilePath string, debugLogFn func(string, ...int
 	return ignorer
 }
 
-// FilterErrors takes a slice of errors and returns a new slice containing only the
-// errors that do not match this Ignorer's ignore string patterns.
-func (i *Ignorer) FilterErrors(errors []error) []error {
-	keepers := make([]error, 0)
-	for _, err := range errors {
-		errText := err.Error()
-		if !i.IsIgnorable(errText) {
-			i.debugFnOverride("not filtering this error because it is not ignorable: %s", errText)
-			keepers = append(keepers, err)
-		}
-		i.debugFnOverride("filtering this error because it is ignorable: %s", errText)
-	}
-
-	return keepers
-}
-
-// FilterMessages takes a slice of support.Message and returns a new slice
-// containing only the Messages that do not match this Ignorer's ignore string patterns.
-func (i *Ignorer) FilterMessages(messages []support.Message) []support.Message {
-	keepers := make([]support.Message, 0)
-	for _, msg := range messages {
-		if !i.IsIgnorable(msg.Err.Error()) {
-			keepers = append(keepers, msg)
-		}
-	}
-	return keepers
-}
-
 // IsIgnoredPathlessError checks a given string to determine whether it looks like a
 // helm lint finding that does not specifically specify an offending file path.
 // These will usually be related to Chart.yaml contents rather than a template
 // inside the chart itself.
-func (i *Ignorer) IsIgnoredPathlessError(errText string) bool {
+func (i *CmdIgnorer) IsIgnoredPathlessError(errText string) bool {
 	for ignorableError := range i.ErrorPatterns {
 		parts := strings.SplitN(ignorableError, ":", 2)
 		prefix := strings.TrimSpace(parts[0])
@@ -91,12 +62,12 @@ func (i *Ignorer) IsIgnoredPathlessError(errText string) bool {
 	return false
 }
 
-// Debug provides an Ignorer with a caller-overridable logging function
+// Debug provides an CmdIgnorer with a caller-overridable logging function
 // intended to match the behavior of the top level debug() method from package main.
 //
 // When no i.debugFnOverride is present Debug will fall back to a naive
 // implementation that assumes all debug output should be logged and not swallowed.
-func (i *Ignorer) Debug(format string, args ...interface{}) {
+func (i *CmdIgnorer) Debug(format string, args ...interface{}) {
 	if i.debugFnOverride == nil {
 		i.debugFnOverride = func(format string, v ...interface{}) {
 			format = fmt.Sprintf("[debug] %s\n", format)
@@ -107,7 +78,7 @@ func (i *Ignorer) Debug(format string, args ...interface{}) {
 	i.debugFnOverride(format, args...)
 }
 
-func (i *Ignorer) IsIgnorable(errText string) bool {
+func (i *CmdIgnorer) IsIgnorable(errText string) bool {
 	errorFullPath, err := extractFullPathFromError(errText)
 	if err != nil {
 		i.Debug("Unable to find a path for error, guess we'll keep it: %s, %v", errText, err)
@@ -145,7 +116,7 @@ func extractFullPathFromError(errText string) (string, error) {
 	return "", fmt.Errorf("fewer than three [%s]-delimited parts found, no path here: %s", delimiter, errText)
 }
 
-func (i *Ignorer) loadFromFilePath(filePath string) {
+func (i *CmdIgnorer) loadFromFilePath(filePath string) {
 	file, err := os.Open(filePath)
 	if err != nil {
 		i.Debug("failed to open lint ignore file: %s", filePath)
@@ -155,7 +126,7 @@ func (i *Ignorer) loadFromFilePath(filePath string) {
 	i.LoadFromReader(file)
 }
 
-func (i *Ignorer) LoadFromReader(rdr io.Reader) {
+func (i *CmdIgnorer) LoadFromReader(rdr io.Reader) {
 	if i.Patterns == nil {
 		i.Patterns = make(map[string][]string)
 	}
